+++ types/array/binarySearch.d.ts
@@ -1,0 +1,1 @@
+export { binarySearch, binarySearch as default } from "../array.js";