+++ types/array/keyBy.d.ts
@@ -1,0 +1,1 @@
+export { keyBy, keyBy as default } from "../array.js";