+++ types/array/lowerBound.d.ts
@@ -1,0 +1,1 @@
+export { lowerBound, lowerBound as default } from "../array.js";