+++ types/array/upperBound.d.ts
@@ -1,0 +1,1 @@
+export { upperBound, upperBound as default } from "../array.js";